# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档",
        "title": "github.com/x-tardis/go-admin API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/wenjianzhang/github.com/x-tardis/go-admin/blob/master/LICENSE.md"
        },
        "version": "1.0.1"
    },
    "paths": {
...
        "/api/v1/dict/type": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "字典类型"
                ],
                "summary": "修改字典类型",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DictType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "字典类型"
                ],
                "summary": "添加字典类型",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DictType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/dict/type/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "字典类型"
                ],
                "summary": "字典类型列表数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dictName",
                        "name": "dictName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictId",
                        "name": "dictId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictType",
                        "name": "dictType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页条数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/paginator.Page"
                        }
                    }
                }
            }
        },
        "/api/v1/dict/type/{dictId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "字典类型"
                ],
                "summary": "通过字典id获取字典类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型编码",
                        "name": "dictId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/servers.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除数据",
                "tags": [
                    "字典类型"
                ],
                "summary": "删除字典类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "dictId",
                        "name": "dictId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.DictType": {
            "type": "object",
            "properties": {
                "createBy": {
                    "description": "创建者",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dataScope": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "dictId": {
                    "type": "integer"
                },
                "dictName": {
                    "description": "字典名称",
                    "type": "string"
                },
                "dictType": {
                    "description": "字典类型",
                    "type": "string"
                },
                "params": {
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "updateBy": {
                    "description": "更新者",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "paginator.Page": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "list": {
                    "type": "object"
                },
                "pageIndex": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
        "servers.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}